* src/pic/gen.c (bitpatternFromVal): fixed for 64-bit machines * src/pic16/ralloc.c (newReg): use correct name in hashtable,   (allocReg): add only new registers to dynAllocRegs,   (pic16_assignRegisters): do not reset dynrIdx, fixes bugs     #1489055, #1445850, and probably #1483693
git-svn-id: https://sdcc.svn.sourceforge.net/svnroot/sdcc/trunk/sdcc@4182 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ src/pic/gen.c
@@ -34,6 +34,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include <ctype.h>
+#include <stdint.h>
 #include "SDCCglobl.h"
 #include "newalloc.h"
 
@@ -9361,10 +9362,10 @@ bitpatternFromVal (value *val)
 {
   union {
     float d;
-    unsigned long l;
+    uint32_t l;
   } float_long;
 
-  assert (sizeof (float) == sizeof (long));
+  assert (sizeof (float) == sizeof (uint32_t));
 
   //fprintf (stderr, "%s:%u(%s): val=%lf, type: %d, etype: %d\n", __FILE__, __LINE__, __FUNCTION__, floatFromVal(val), SPEC_NOUN(val->type), SPEC_NOUN(val->etype));
